Warning: Possibly disruptive changes.
Added three new widget areas (below header, below single post, page footer).

Number of columns is now based on whether any widgets are present in the sidebars, which removes the need for a column option on the K2 Options page.

Also started cleaning up the CSS and structure a little. More to come.

I  was planning on this for 1.2, but then I accidentally finished it today.